requirements.txt
python-telegram-bot>=20.0
gigachat
nest-asyncio
SpeechRecognition
pydub
python-dotenv
openai
